mcp_server/requirements.txt
mcp>=1.9.4
ollama>=0.1.0
redis>=5.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
